finished moving all private APIs to internal dir, still need to get FSM tests working
